Avoid PHP scalar type juggling in includes/ (part 2)
Continuation of e5444ea55a8000f0040.

+++ includes/auth/LocalPasswordPrimaryAuthenticationProvider.php
@@ -54,14 +54,14 @@ class LocalPasswordPrimaryAuthenticationProvider
 	 * @return \stdClass|null
 	 */
 	protected function getPasswordResetData( $username, $row ) {
-		$now = wfTimestamp();
+		$now = (int)wfTimestamp();
 		$expiration = wfTimestampOrNull( TS_UNIX, $row->user_password_expires );
-		if ( $expiration === null || $expiration >= $now ) {
+		if ( $expiration === null || (int)$expiration >= $now ) {
 			return null;
 		}
 
 		$grace = $this->config->get( 'PasswordExpireGrace' );
-		if ( $expiration + $grace < $now ) {
+		if ( (int)$expiration + $grace < $now ) {
 			$data = [
 				'hard' => true,
 				'msg' => \Status::newFatal( 'resetpass-expired' )->getMessage(),
